doc: Hookup every exported header into groups and change main page
Also merge sol-blob.h with sol-types.h -- there was no reason it got an
exclusive header.
